Move DB files to database folder for efficiency
